+++ jsonargparse_tests/test_typehints.py
@@ -67,6 +67,12 @@ def test_add_argument_failure_given_type_and_action(parser):
     ctx.match("Providing both type and action not allowed")
 
 
+@pytest.mark.parametrize("typehint", [Namespace, Optional[Namespace], Union[int, Namespace], List[Namespace]])
+def test_namespace_unsupported_as_type(parser, typehint):
+    with pytest.raises(ValueError, match="Namespace .* not supported as a type"):
+        parser.add_argument("--ns", type=typehint)
+
+
 # basic types tests